+++ src/main/java/cc/cassian/immersiveoverlays/compat/ModCompat.java
@@ -1,8 +1,9 @@
-package cc.cassian.immersiveoverlays;
+package cc.cassian.immersiveoverlays.compat;
 
 import cc.cassian.immersiveoverlays.helpers.ModHelpers;
 
 public class ModCompat {
+    public static final boolean ACCESSORIES = ModHelpers.isLoaded("accessories");
     public static final boolean MAP_ATLASES = ModHelpers.isLoaded("map_atlases");
     public static final boolean SPELUNKERY = ModHelpers.isLoaded("spelunkery");
     public static final boolean SUPPLEMENTARIES = ModHelpers.isLoaded("supplementaries");