+++ src/main/java/cc/cassian/immersiveoverlays/overlay/BiomeOverlay.java
@@ -3,21 +3,14 @@ package cc.cassian.immersiveoverlays.overlay;
 import cc.cassian.immersiveoverlays.config.ModConfig;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
-import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.player.LocalPlayer;
-import net.minecraft.client.renderer.texture.MissingTextureAtlasSprite;
-import net.minecraft.client.renderer.texture.TextureAtlas;
 import net.minecraft.client.resources.language.I18n;
-import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
-import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.level.biome.Biome;
 import net.minecraft.world.level.biome.Biomes;
 import org.apache.commons.lang3.text.WordUtils;
 
-import static cc.cassian.immersiveoverlays.overlay.OverlayHelpers.TEXTURE;
+import static cc.cassian.immersiveoverlays.ModClient.MOD_ID;
 
 public class BiomeOverlay {
     public static boolean showBiome = true;
@@ -34,7 +27,8 @@ public class BiomeOverlay {
             return;
         var mc = Minecraft.getInstance();
 
-        var time = getBiome(mc.player);
+        ResourceLocation biome = getBiome(mc.player);
+        String biomeText = formatBiome(biome);
 
         int xOffset = 3;
         // The amount of offset needed to display the biome icons.
@@ -44,7 +38,7 @@ public class BiomeOverlay {
         int yPlacement = ModConfig.get().biome_vertical_position;
         int textYPlacement = yPlacement+2;
 
-        int fontWidth = mc.font.width(time)+iconOffset;
+        int fontWidth = mc.font.width(biomeText)+iconOffset;
 
         if (mc.player == null) return;
         if (OverlayHelpers.playerHasPotions(mc.player)) {
@@ -59,42 +53,48 @@ public class BiomeOverlay {
         OverlayHelpers.renderBackground(poseStack, windowWidth, fontWidth, xPlacement, xOffset, yPlacement, textureOffset, tooltipSize);
         // render text
         //? if >1.20 {
-        poseStack.drawString(mc.font, time, xPlacement-xOffset+iconOffset, textYPlacement, 14737632);
+        poseStack.drawString(mc.font, biomeText, xPlacement-xOffset+iconOffset, textYPlacement, 14737632);
         //?} else {
         /*GuiComponent.drawString(poseStack, mc.font, time, xPlacement-xOffset+iconOffset, textYPlacement, 14737632);
          *///?}
-        var spriteOffset = getBiomeOffset(mc.player);
+        var sprite = getBiomeSprite(biome);
 
         //? if >1.21.2 {
-        /*poseStack.blit(RenderType::guiTextured, TEXTURE,
+        /*poseStack.blit(RenderType::guiTextured, sprite,
         *///?} else if >1.20 {
-        poseStack.blit(TEXTURE,
+        poseStack.blit(sprite,
         //?} else {
 
-        /*RenderSystem.setShaderTexture(0, TEXTURE);
+        /*RenderSystem.setShaderTexture(0, sprite);
            GuiComponent.blit(poseStack,
          *///?}
-                xPlacement-xOffset-1, yPlacement-1,
-                0, spriteOffset,
-                95, 16, 16,
-                OverlayHelpers.textureSize, OverlayHelpers.textureSize);
+                xPlacement-xOffset-1, yPlacement-2,
+                0, 0,
+                0, 16, 16,
+                16, 16);
     }
 
-    private static int getBiomeOffset(LocalPlayer player) {
-//        Minecraft.getInstance().getResourceManager().getResource(TEXTURE);
-        return 0;
+    private static ResourceLocation getBiomeSprite(ResourceLocation biome) {
+        var key = ResourceLocation.tryBuild(biome.getNamespace(), "textures/immersiveoverlays/"+ biome.getPath() +".png");
+        if (Minecraft.getInstance().getResourceManager().getResource(key).isPresent())
+            return key;
+        else
+            return ResourceLocation.tryBuild(MOD_ID, "textures/immersiveoverlays/undefined.png");
     }
 
-    private static String getBiome(LocalPlayer player) {
+    private static ResourceLocation getBiome(LocalPlayer player) {
         //? if >1.20 {
         var level = player.level();
         //?} else {
         /*var level = player.level;
          *///?}
-        ResourceLocation biome = level.getBiome(player.blockPosition()).unwrapKey().orElse(Biomes.THE_VOID).location();
+        return level.getBiome(player.blockPosition()).unwrapKey().orElse(Biomes.THE_VOID).location();
+    }
+
+    public static String formatBiome(ResourceLocation biome) {
         var path = biome.getPath();
         var namespace = biome.getNamespace();
-        var key = "biome." + namespace + path;
+        var key = "biome.%s.%s".formatted(namespace, path);
         if (I18n.exists(key)) {
             return I18n.get(key);
         } else return WordUtils.capitalize(path);